+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `lang-util`: Add NodeSpan::start() and NodeSpan::end()
 - `lang-util`: Generate display impl for derive(Token)
 - `lang-util`: Add lang_util(as = "...") for mapping to lalrpop parser tokens
+- `lang-util`: Add TokenDescription metadata for lexer tokens
 
 ### Changed
 - PartialEq for Node now compares the contents of the node by default